third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/attr-dir-inherit-expected.txt

CONSOLE ERROR: line 29: Uncaught TypeError: parent.createShadowRoot is not a function
This is a testharness.js-based test.
FAIL attr-dir-inherit Uncaught TypeError: parent.createShadowRoot is not a function
Harness: the test ran to completion.
